refactor: organize `keyboard`
