+++ apps/calendar/src/util/keyboard.ts
@@ -1,4 +1,4 @@
-import { ESCAPE, KEYCODE } from '@src/util/constants';
+import { ESCAPE, KEYCODE } from '@src/constants/keyboard';
 
 export function isEscapePressed(e: KeyboardEvent) {
   return e.key ? e.key === ESCAPE : e.keyCode === KEYCODE[ESCAPE];
